Arrange/Split Null tests files
