+++ src/primitives/Primitives.Tests/PredicateExtensions.Tests/IsNull_Anonymous.cs
@@ -1,0 +1,23 @@
+﻿#nullable enable
+
+using NUnit.Framework;
+using System;
+using static PrimeFuncPack.UnitTest.TestData;
+
+namespace PrimeFuncPack.Primitives.Tests
+{
+    partial class PredicateExtensionsTests
+    {
+        [Test]
+        public void IsNull_AnonymousValueIsNotNull_ExpectFalse()
+        {
+            var source = new
+            {
+                Text = SomeTextStructType
+            };
+
+            var actual = source.IsNull();
+            Assert.False(actual);
+        }
+    }
+}
