+++ src/primitives/Primitives.Tests/PredicateExtensions.Tests/IsNull_NullableStruct.cs
@@ -9,36 +9,6 @@ namespace PrimeFuncPack.Primitives.Tests
     partial class PredicateExtensionsTests
     {
         [Test]
-        public void IsNull_RefValueIsNull_ExpectTrue()
-        {
-            object source = null!;
-
-            var actual = source.IsNull();
-            Assert.True(actual);
-        }
-
-        [Test]
-        public void IsNull_RefValueIsNotNull_ExpectFalse()
-        {
-            var source = ZeroIdRefType;
-
-            var actual = source.IsNull();
-            Assert.False(actual);
-        }
-
-        [Test]
-        public void IsNull_AnonimousValueIsNotNull_ExpectFalse()
-        {
-            var source = new
-            {
-                Text = SomeTextStructType
-            };
-
-            var actual = source.IsNull();
-            Assert.False(actual);
-        }
-
-        [Test]
         public void IsNull_NullableStructIsNull_CallStructOverload_ExpectTrue()
         {
             int? source = null;
@@ -73,16 +43,6 @@ namespace PrimeFuncPack.Primitives.Tests
         public void IsNull_NullableStructIsNotNull_CallObjectOverload_ExpectFalse(int? source)
         {
             var actual = source.IsNull<object>();
-            Assert.False(actual);
-        }
-
-        // Test both zero and non-zero values
-        [Test]
-        [TestCase(0)]
-        [TestCase(1)]
-        public void IsNull_NonnullableStruct_ExpectFalse(int source)
-        {
-            var actual = source.IsNull();
             Assert.False(actual);
         }
     }
