+++ src/primitives/Primitives.Tests/Predicates.Tests/IsNotNull_NullableStruct.cs
@@ -2,42 +2,11 @@
 
 using NUnit.Framework;
 using System;
-using static PrimeFuncPack.UnitTest.TestData;
 
 namespace PrimeFuncPack.Primitives.Tests
 {
     public sealed partial class PredicatesTests
     {
-        [Test]
-        public void IsNotNull_RefValueIsNull_ExpectFalse()
-        {
-            object source = null!;
-
-            var actual = Predicates.IsNotNull(source);
-            Assert.False(actual);
-        }
-
-        [Test]
-        public void IsNotNull_RefValueIsNotNull_ExpectTrue()
-        {
-            var source = PlusFifteenIdRefType;
-
-            var actual = Predicates.IsNotNull(source);
-            Assert.True(actual);
-        }
-
-        [Test]
-        public void IsNotNull_AnonimousValueIsNotNull_ExpectTrue()
-        {
-            var source = new
-            {
-                Text = SomeTextStructType
-            };
-
-            var actual = Predicates.IsNotNull(source);
-            Assert.True(actual);
-        }
-
         [Test]
         public void IsNotNull_NullableStructIsNull_CallStructOverload_ExpectFalse()
         {
@@ -73,16 +42,6 @@ namespace PrimeFuncPack.Primitives.Tests
         public void IsNotNull_NullableStructIsNotNull_CallObjectOverload_ExpectTrue(int? source)
         {
             var actual = Predicates.IsNotNull<object>(source);
-            Assert.True(actual);
-        }
-
-        // Test both zero and non-zero values
-        [Test]
-        [TestCase(0)]
-        [TestCase(1)]
-        public void IsNotNull_NonnullableStruct_ExpectTrue(int source)
-        {
-            var actual = Predicates.IsNotNull(source);
             Assert.True(actual);
         }
     }
