+++ src/primitives/Primitives.Tests/Predicates.Tests/IsNull_Anonymous.cs
@@ -1,0 +1,23 @@
+﻿#nullable enable
+
+using NUnit.Framework;
+using System;
+using static PrimeFuncPack.UnitTest.TestData;
+
+namespace PrimeFuncPack.Primitives.Tests
+{
+    partial class PredicatesTests
+    {
+        [Test]
+        public void IsNull_AnonymousValueIsNotNull_ExpectFalse()
+        {
+            var source = new
+            {
+                Text = SomeTextStructType
+            };
+
+            var actual = Predicates.IsNull(source);
+            Assert.False(actual);
+        }
+    }
+}
