+++ src/primitives/Primitives.Tests/Predicates.Tests/IsNull_NullableStruct.cs
@@ -1,44 +1,12 @@
 ﻿#nullable enable
 
 using NUnit.Framework;
-using PrimeFuncPack.UnitTest;
 using System;
-using static PrimeFuncPack.UnitTest.TestData;
 
 namespace PrimeFuncPack.Primitives.Tests
 {
     partial class PredicatesTests
     {
-        [Test]
-        public void IsNull_RefValueIsNull_ExpectTrue()
-        {
-            RefType source = null!;
-
-            var actual = Predicates.IsNull(source);
-            Assert.True(actual);
-        }
-
-        [Test]
-        public void IsNull_RefValueIsNull_ExpectFalse()
-        {
-            var source = MinusFifteenIdRefType;
-
-            var actual = Predicates.IsNull(source);
-            Assert.False(actual);
-        }
-
-        [Test]
-        public void IsNull_AnonimousValueIsNotNull_ExpectFalse()
-        {
-            var source = new
-            {
-                Text = SomeTextStructType
-            };
-
-            var actual = Predicates.IsNull(source);
-            Assert.False(actual);
-        }
-
         [Test]
         public void IsNull_NullableStructIsNull_CallStructOverload_ExpectTrue()
         {
@@ -74,16 +42,6 @@ namespace PrimeFuncPack.Primitives.Tests
         public void IsNull_NullableStructIsNotNull_CallObjectOverload_ExpectFalse(int? source)
         {
             var actual = Predicates.IsNull<object>(source);
-            Assert.False(actual);
-        }
-
-        // Test both zero and non-zero values
-        [Test]
-        [TestCase(0)]
-        [TestCase(1)]
-        public void IsNull_NonnullableStruct_ExpectFalse(int source)
-        {
-            var actual = Predicates.IsNull(source);
             Assert.False(actual);
         }
     }
